refactor: split tax advantaged category API module by responsibility
