+++ frontend/src/api/taxAdvantagedCategories/index.ts
@@ -1,0 +1,33 @@
+export type {
+  CreateTaxAdvantagedCategoryLimitPayload,
+  CreateTaxAdvantagedCategoryPayload,
+  TaxAdvantagedCategory,
+  TaxAdvantagedCategoryLimit,
+  TaxTreatment,
+  UpdateTaxAdvantagedCategoryLimitPayload,
+  UpdateTaxAdvantagedCategoryPayload,
+} from '@/api/taxAdvantagedCategories/types';
+
+export {
+  createTaxAdvantagedCategory,
+  createTaxAdvantagedCategoryLimit,
+  deleteTaxAdvantagedCategory,
+  deleteTaxAdvantagedCategoryLimit,
+  fetchTaxAdvantagedCategories,
+  fetchTaxAdvantagedCategory,
+  fetchTaxAdvantagedCategoryLimits,
+  updateTaxAdvantagedCategory,
+  updateTaxAdvantagedCategoryLimit,
+} from '@/api/taxAdvantagedCategories/requests';
+
+export {
+  useCreateTaxAdvantagedCategory,
+  useCreateTaxAdvantagedCategoryLimit,
+  useDeleteTaxAdvantagedCategory,
+  useDeleteTaxAdvantagedCategoryLimit,
+  useTaxAdvantagedCategories,
+  useTaxAdvantagedCategory,
+  useTaxAdvantagedCategoryLimits,
+  useUpdateTaxAdvantagedCategory,
+  useUpdateTaxAdvantagedCategoryLimit,
+} from '@/api/taxAdvantagedCategories/hooks';
